Move notebook export to overlay
